feat(7): wip refactor and add tests
