+++ core/Cargo.toml
@@ -23,6 +23,7 @@ zeroize = "1.5.7"
 pretty_assertions = "1.3.0"
 rand = "0.8.5"
 hex = "0.4.3"
+base64 = "*"
 
 [features]
 wasm = ["wasm-bindgen"]
