+++ core/src/api/api.rs
@@ -1,22 +1,35 @@
-use crate::dsnp::{
-	api_types::{
-		Connection, ConnectionType, DsnpKeys, ExportBundle, ImportBundle, PrivacyType, PublicKey,
+use crate::{
+	dsnp::{
+		api_types::{
+			Connection, ConnectionType, DsnpKeys, ExportBundle, ImportBundle, PrivacyType,
+			PublicKey,
+		},
+		dsnp_types::{DsnpGraphEdge, DsnpUserId},
+		encryption::EncryptionBehavior,
 	},
-	dsnp_types::DsnpUserId,
-	encryption::EncryptionBehavior,
-	graph_page::UserGraph,
+	graph::{updates::UpdateEvent, user::UserGraph},
+	iter_graph_connections,
+	util::time::time_in_ksecs,
 };
 use anyhow::{Error, Result};
-use std::{cmp::min, collections::HashMap};
+use std::{cmp::min, collections::HashMap, marker::PhantomData};
 
 const MAX_GRAPH_USERS_DEFAULT: usize = 1000;
 
 #[allow(non_snake_case)]
-pub struct GraphState<const MAX_USERS: usize = MAX_GRAPH_USERS_DEFAULT> {
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub struct GraphState<E: EncryptionBehavior, const MAX_USERS: usize = MAX_GRAPH_USERS_DEFAULT> {
+	_encryption: PhantomData<E>,
 	user_map: HashMap<DsnpUserId, UserGraph>,
 }
 
 pub trait GraphAPI<E: EncryptionBehavior> {
+	/// Check if graph state contains a user
+	fn contains_user(&self, user_id: &DsnpUserId) -> bool;
+
+	/// Return number of users in the current graph state
+	fn len(&self) -> usize;
+
 	/// Create a new, empty user graph
 	fn add_user_graph(&mut self, user_id: &DsnpUserId) -> Result<&mut UserGraph>;
 
@@ -38,31 +51,57 @@ pub trait GraphAPI<E: EncryptionBehavior> {
 	) -> Result<Vec<ExportBundle>>;
 
 	/// Add the connection to the list of pending additions for the user
-	fn add_connection_for_user(user_id: &DsnpUserId, connection: &Connection) -> Result<()>;
+	fn add_connection_for_user(
+		&mut self,
+		user_id: &DsnpUserId,
+		connection: &Connection,
+	) -> Result<()>;
 
 	/// Add an indication that the connection is pending removal for the user
-	fn remove_connection_from_user(user_id: &DsnpUserId, connection: &Connection);
+	fn remove_connection_from_user(
+		&mut self,
+		user_id: &DsnpUserId,
+		connection: &Connection,
+	) -> Result<()>;
 
 	/// Get a list of all connections of the indicated type for the user
 	fn get_connections_for_user_graph(
+		&self,
 		user_id: &DsnpUserId,
 		connection_type: &ConnectionType,
 		include_pending: bool,
-	) -> Vec<Connection>;
+	) -> Result<Vec<DsnpGraphEdge>>;
 }
 
-impl<const MAX_USERS: usize> GraphState<MAX_USERS> {
+impl<const MAX_USERS: usize, E: EncryptionBehavior> GraphState<E, MAX_USERS> {
 	pub fn new() -> Self {
-		Self { user_map: HashMap::<DsnpUserId, UserGraph>::new() }
+		Self { _encryption: PhantomData, user_map: HashMap::<DsnpUserId, UserGraph>::new() }
 	}
 
 	pub fn with_capacity(capacity: usize) -> Self {
 		let size = min(capacity, MAX_USERS);
-		Self { user_map: HashMap::<DsnpUserId, UserGraph>::with_capacity(size) }
+		Self {
+			_encryption: PhantomData,
+			user_map: HashMap::<DsnpUserId, UserGraph>::with_capacity(size),
+		}
+	}
+
+	pub fn capacity(&self) -> usize {
+		MAX_USERS
 	}
 }
 
-impl<E: EncryptionBehavior, const M: usize> GraphAPI<E> for GraphState<M> {
+impl<E: EncryptionBehavior, const M: usize> GraphAPI<E> for GraphState<E, M> {
+	/// Check if graph state contains a user
+	fn contains_user(&self, user_id: &DsnpUserId) -> bool {
+		self.user_map.contains_key(user_id)
+	}
+
+	/// Return number of users in the current graph state
+	fn len(&self) -> usize {
+		self.user_map.len()
+	}
+
 	/// Create a new, empty user graph
 	fn add_user_graph(&mut self, user_id: &DsnpUserId) -> Result<&mut UserGraph> {
 		if self.user_map.len() >= M {
@@ -93,7 +132,7 @@ impl<E: EncryptionBehavior, const M: usize> GraphAPI<E> for GraphState<M> {
 	fn import_user_data(&mut self, payload: ImportBundle<E>) -> Result<()> {
 		let user_graph = match self.user_map.get_mut(&payload.dsnp_user_id) {
 			Some(graph) => graph,
-			None => <GraphState<M> as GraphAPI<E>>::add_user_graph(self, &payload.dsnp_user_id)?,
+			None => self.add_user_graph(&payload.dsnp_user_id)?,
 		};
 
 		let graph = user_graph.graph_mut(&payload.connection_type);
@@ -130,21 +169,235 @@ impl<E: EncryptionBehavior, const M: usize> GraphAPI<E> for GraphState<M> {
 	}
 
 	/// Add the connection to the list of pending additions for the user
-	fn add_connection_for_user(_user_id: &DsnpUserId, _connection: &Connection) -> Result<()> {
-		todo!();
+	fn add_connection_for_user(
+		&mut self,
+		user_id: &DsnpUserId,
+		connection: &Connection,
+	) -> Result<()> {
+		let user_graph = match self.user_map.get_mut(user_id) {
+			Some(g) => g,
+			None => return Err(Error::msg("user graph not found in state")),
+		};
+
+		user_graph.update_tracker.register_update(&UpdateEvent::create_add(
+			connection.dsnp_user_id,
+			connection.connection_type,
+		))
 	}
 
 	/// Add an indication that the connection is pending removal for the user
-	fn remove_connection_from_user(_user_id: &DsnpUserId, _connection: &Connection) {
-		todo!();
+	fn remove_connection_from_user(
+		&mut self,
+		user_id: &DsnpUserId,
+		connection: &Connection,
+	) -> Result<()> {
+		let graph = match self.user_map.get_mut(user_id) {
+			Some(g) => g,
+			None => return Err(Error::msg("user graph not found in state")),
+		};
+		graph.update_tracker.register_update(&UpdateEvent::create_remove(
+			connection.dsnp_user_id,
+			connection.connection_type,
+		))
 	}
 
 	/// Get a list of all connections of the indicated type for the user
 	fn get_connections_for_user_graph(
-		_user_id: &DsnpUserId,
-		_connection_type: &ConnectionType,
-		_include_pending: bool,
-	) -> Vec<Connection> {
-		todo!();
+		&self,
+		user_id: &DsnpUserId,
+		connection_type: &ConnectionType,
+		include_pending: bool,
+	) -> Result<Vec<DsnpGraphEdge>> {
+		let user_graph = match self.user_map.get(user_id) {
+			Some(graph) => graph,
+			None => return Err(Error::msg("user not present in graph state")),
+		};
+
+		let graph = user_graph.graph(connection_type);
+		let mut connections: Vec<DsnpGraphEdge> = iter_graph_connections!(graph).cloned().collect();
+
+		if include_pending {
+			user_graph
+				.update_tracker
+				.get_updates_for_connection_type(*connection_type)
+				.unwrap_or(&Vec::<UpdateEvent>::new())
+				.iter()
+				.cloned()
+				.for_each(|event| match event {
+					UpdateEvent::Add { dsnp_user_id, .. } =>
+						if !connections.iter().map(|c| c.user_id).any(|id| id == dsnp_user_id) {
+							connections.push(DsnpGraphEdge {
+								user_id: dsnp_user_id,
+								since: time_in_ksecs(),
+							})
+						},
+					UpdateEvent::Remove { dsnp_user_id, .. } =>
+						connections.retain(|c| c.user_id != dsnp_user_id),
+				});
+		}
+
+		Ok(connections)
 	}
+}
+
+#[cfg(test)]
+mod test {
+	use crate::dsnp::encryption::SealBox;
+
+	use super::*;
+	const TEST_CAPACITY: usize = 10;
+
+	type TestGraphState<const M: usize = TEST_CAPACITY> = GraphState<SealBox, M>;
+
+	#[test]
+	fn new_graph_state_with_capacity_sets_initial_hash_map_capacity() {
+		let capacity: usize = 5;
+		let new_state = TestGraphState::<TEST_CAPACITY>::with_capacity(capacity);
+		assert!(new_state.user_map.capacity() >= capacity);
+	}
+
+	#[test]
+	fn new_graph_state_with_capacity_caps_initial_hash_map_capacity() {
+		let new_state = TestGraphState::<TEST_CAPACITY>::with_capacity(TEST_CAPACITY * 2);
+		assert!(new_state.user_map.capacity() >= TEST_CAPACITY);
+	}
+
+	#[test]
+	fn graph_state_capacity() {
+		let state: TestGraphState = TestGraphState::new();
+		assert_eq!(state.capacity(), TEST_CAPACITY);
+	}
+
+	#[test]
+	fn graph_contains_false() {
+		let state: TestGraphState = TestGraphState::new();
+		assert!(!state.contains_user(&0));
+	}
+
+	#[test]
+	fn graph_contains_true() {
+		let mut state: TestGraphState = TestGraphState::new();
+		let _ = state.add_user_graph(&0);
+		assert!(state.contains_user(&0));
+	}
+
+	#[test]
+	fn graph_len() {
+		let mut state: TestGraphState = TestGraphState::new();
+		let _ = state.add_user_graph(&0);
+		assert_eq!(state.len(), 1);
+		let _ = state.add_user_graph(&1);
+		assert_eq!(state.len(), 2);
+	}
+
+	#[test]
+	fn add_user_errors_if_graph_state_full() {
+		let mut state = TestGraphState::<1>::new();
+		let _ = state.add_user_graph(&0);
+		assert!(state.add_user_graph(&1).is_err());
+	}
+
+	#[test]
+	fn add_duplicate_user_errors() {
+		let mut state: TestGraphState = TestGraphState::new();
+		let _ = state.add_user_graph(&0);
+		assert!(state.add_user_graph(&0).is_err());
+	}
+	#[test]
+	fn add_user_success() -> Result<()> {
+		let mut state: TestGraphState = TestGraphState::new();
+
+		state.add_user_graph(&0)?;
+		Ok(())
+	}
+
+	#[test]
+	fn remove_user_success() {
+		let mut state: TestGraphState = TestGraphState::new();
+		let _ = state.add_user_graph(&0);
+		let _ = state.add_user_graph(&1);
+		state.remove_user_graph(&0);
+		assert_eq!(state.len(), 1);
+		assert!(!state.contains_user(&0));
+		assert!(state.contains_user(&1));
+	}
+
+	#[test]
+	fn remove_nonexistent_user_noop() {
+		let mut state: TestGraphState = TestGraphState::new();
+		let _ = state.add_user_graph(&0);
+		let _ = state.add_user_graph(&1);
+		let state_copy = state.clone();
+		state.remove_user_graph(&99);
+		assert_eq!(state, state_copy);
+	}
+
+	#[test]
+	#[ignore = "todo"]
+	fn import_user_data() {}
+
+	#[test]
+	#[ignore = "todo"]
+	fn export_user_updates() {}
+
+	#[test]
+	fn add_duplicate_connection_for_user_errors() {
+		let connection = Connection {
+			connection_type: ConnectionType::Follow(PrivacyType::Private),
+			dsnp_user_id: 1,
+		};
+
+		let mut state: TestGraphState = TestGraphState::new();
+		let _ = state.add_user_graph(&0);
+		assert!(state.add_connection_for_user(&0, &connection).is_ok());
+		assert!(state.add_connection_for_user(&0, &connection).is_err());
+	}
+
+	#[test]
+	fn add_connection_for_nonexistent_user_errors() {
+		let mut state: TestGraphState = TestGraphState::new();
+		assert!(state
+			.add_connection_for_user(
+				&0,
+				&Connection {
+					dsnp_user_id: 1,
+					connection_type: ConnectionType::Follow(PrivacyType::Private),
+				},
+			)
+			.is_err());
+	}
+
+	#[test]
+	fn remove_connection_for_user_twice_errors() {
+		let connection: Connection = Connection {
+			dsnp_user_id: 1,
+			connection_type: ConnectionType::Follow(PrivacyType::Private),
+		};
+		let mut state: TestGraphState = TestGraphState::new();
+		let _ = state.add_user_graph(&0);
+		assert!(state.remove_connection_from_user(&0, &connection).is_ok());
+		assert!(state.remove_connection_from_user(&0, &connection).is_err());
+	}
+
+	#[test]
+	fn remove_connection_from_nonexistent_user_errors() {
+		let mut state: TestGraphState = TestGraphState::new();
+		assert!(state
+			.remove_connection_from_user(
+				&0,
+				&Connection {
+					dsnp_user_id: 1,
+					connection_type: ConnectionType::Follow(PrivacyType::Private),
+				}
+			)
+			.is_err());
+	}
+
+	#[test]
+	#[ignore = "todo"]
+	fn get_connections_for_user_graph() {}
+
+	#[test]
+	#[ignore = "todo"]
+	fn get_connections_for_user_graph_with_pending() {}
 }
