+++ core/src/dsnp/mod.rs
@@ -2,7 +2,6 @@ pub mod api_types;
 pub mod compression;
 pub mod dsnp_types;
 pub mod encryption;
-pub mod graph_page;
 mod pseudo_relationship_identifier;
 pub mod reader_writer;
 pub mod schema;
