+++ core/src/dsnp/tests/mod.rs
@@ -1,7 +1,5 @@
 mod compression;
 mod dsnp_types;
 mod encryption;
-mod graph_page;
-mod helpers;
 mod pseudo_relationship_identifier;
 mod schema;
