+++ core/src/frequency/tests/mod.rs
@@ -1,2 +1,1 @@
-mod config;
 mod reader_writer;
