+++ core/src/graph/mod.rs
@@ -1,1 +1,6 @@
+pub mod graph;
+#[cfg(test)]
+mod helpers;
+pub mod page;
 pub mod updates;
+pub mod user;
